fixed server error handler, refactored test code
