+++ tests/test_error_handling.py
@@ -1,27 +1,11 @@
-import pytest
-import json
 from test_plus import APITestCase
 
-from wazimap_ng.utils import error_handler
 from rest_framework import status
 
 
-@pytest.mark.django_db
 class TestErrorResponseData(APITestCase):
 
-    def test_error_json(self):
-        request = self.get('/foo/bar')
-
-        response = error_handler(request)
-        content_type = response['Content-Type']
-
-        data = json.loads(response.content)
-
-        assert content_type == 'application/json'
-        assert response.status_code == status.HTTP_500_INTERNAL_SERVER_ERROR
-        assert data['error']['type'] == 'server_error'
-
-    def test_error_format(self):
+    def test_404_error_format(self):
         expected_response = {
             'error': {
                 "message": {
@@ -34,6 +18,10 @@ class TestErrorResponseData(APITestCase):
 
         response = self.get('dataset-detail', pk=234)
 
-        data = json.loads(response.content)
+        assert response.data == expected_response
 
-        assert data == expected_response
+    def test_custom_exception(self):
+        response = self.get('/api/v1/datasets/3452/')
+
+        assert response.status_code == status.HTTP_404_NOT_FOUND
+        assert response.data['error']['type'] == 'not_found'
